+++ atrain_match/statistics/orrb_CFC_stat.py
@@ -51,9 +51,9 @@ class CloudFractionStats(OrrbStats):
             max_prob[-1] = 100
             Num_cloudy_tot = np.sum(n_cloudy_cmaprob)
             Num_clear_tot = np.sum(n_clear_cmaprob)
-            percent_cloudy_prob = np.array([100.0 / Num_cloudy_tot * np.int(nc) for nc in n_cloudy_cmaprob])
+            percent_cloudy_prob = np.array([100.0 / Num_cloudy_tot * np.int64(nc) for nc in n_cloudy_cmaprob])
             # print(percent_cloudy_prob, Num_cloudy_tot, n_cloudy_cmaprob)
-            percent_clear_prob = np.array([100.0 / Num_clear_tot * np.int(nc) for nc in n_clear_cmaprob])
+            percent_clear_prob = np.array([100.0 / Num_clear_tot * np.int64(nc) for nc in n_clear_cmaprob])
             # print(percent_clear_prob)
 
             detected_clouds = np.array([np.sum(n_cloudy_cmaprob[min_prob >= limit]) for limit in limit_v])
